% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_ranges.R
\name{add_days_before}
\alias{add_days_before}
\title{Add Dates of 1:n Days Before Given Dates}
\usage{
add_days_before(dates, n_days_before = 1)
}
\arguments{
\item{dates}{vector of \code{Date objects}}

\item{n_days_before}{number of days to add before the days given in
\code{dates}}
}
\description{
Add Dates of 1:n Days Before Given Dates
}
\examples{
dates <- as.Date(c("2019-07-05", "2019-07-10"))
fhpredict:::add_days_before(dates)
fhpredict:::add_days_before(dates, n_days_before = 2)

# Duplicates are removed so that unique dates are returned
fhpredict:::add_days_before(dates, n_days_before = 10)
}
