% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.R
\name{api_add_rain}
\alias{api_add_rain}
\title{Add Rain Data Frame to Postgres Database Via API}
\usage{
api_add_rain(user_id, spot_id, rain, time_string = "12:00:00",
  comment = NULL)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{rain}{data frame containing rain data (columns \code{datum},
\code{rain})}

\item{time_string}{time to which rain data values relate, default:
\code{"12:00:00"}}

\item{comment}{character string to be written to the field "comment" of the
rain database table.}
}
\description{
Add Rain Data Frame to Postgres Database Via API
}
\examples{
\dontrun{
# Define some fake rain data
rain <- data.frame(
  datum = as.Date(c("2019-08-29", "2019-08-30")),
  rain = c(1.23, 2.34)
)

# Add the rain data to the database
rain_ids <- api_add_rain(user_id = 3, spot_id = 1441, rain)

Show the IDs of the created database records
rain_ids

# Delete the fake rain data
api_delete_rain(user_id = 3, spot_id = 1441, ids = rain_ids)
}

}
