% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_measurements.R
\name{api_delete_measurements}
\alias{api_delete_measurements}
\title{Delete Discharge Data from Postgres Database Via API}
\usage{
api_delete_measurements(user_id, spot_id, ids = NULL, dbg = TRUE)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{ids}{optional. Vector of measurement ids. If not given or \code{NULL}
(the default) all measurements for the bathing spot are deleted!}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Delete Discharge Data from Postgres Database Via API
}
