% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_model.R
\name{api_delete_model}
\alias{api_delete_model}
\title{Delete a Model from the Database}
\usage{
api_delete_model(user_id, spot_id, model_id)
}
\arguments{
\item{user_id}{user ID}

\item{spot_id}{bathing spot ID}

\item{model_id}{model ID}
}
\description{
This function deletes a model from the database.
}
