% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{api_delete_predictions}
\alias{api_delete_predictions}
\title{Delete Prediction Data from Postgres Database Via API}
\usage{
api_delete_predictions(user_id, spot_id, ids = NULL, dbg = TRUE)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{ids}{optional. Vector of prediction ids. If not given or \code{NULL}
(the default) all predictions for the bathing spot are deleted!}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Delete Prediction Data from Postgres Database Via API
}
