% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.R
\name{api_delete_rain}
\alias{api_delete_rain}
\title{Delete Rain Data from Postgres Database Via API}
\usage{
api_delete_rain(user_id, spot_id, ids = NULL, dbg = TRUE)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{ids}{optional. Vector of rain ids. If not given or \code{NULL} (the
default) all rain data for the bathing spot are deleted!}

\item{dbg}{if \code{TRUE} debug messages are shown}
}
\description{
Delete Rain Data from Postgres Database Via API
}
