% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_accessors.R
\name{api_get_bathingspot}
\alias{api_get_bathingspot}
\title{Get Bathing Spot(s)}
\usage{
api_get_bathingspot(user_id = -1L, spot_id = -1L, pattern = "")
}
\arguments{
\item{user_id}{user ID or -1L (all users)}

\item{spot_id}{bathing spot ID or -1L (all bathing spots)}

\item{pattern}{optional. Pattern matching the names of bathing spots to be
returned}
}
\description{
Get Bathing Spot(s)
}
