% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discharge.R
\name{api_get_discharge}
\alias{api_get_discharge}
\title{Get Discharge Data from the Postgres Database Via API}
\usage{
api_get_discharge(user_id, spot_id)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}
}
\description{
Get Discharge Data from the Postgres Database Via API
}
