% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{api_get_predictions}
\alias{api_get_predictions}
\title{Get Predictions}
\usage{
api_get_predictions(user_id, spot_id)
}
\arguments{
\item{user_id}{user ID or -1L (all users)}

\item{spot_id}{bathing spot ID or -1L (all bathing spots)}
}
\description{
Get Predictions
}
