% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain.R
\name{api_get_rain}
\alias{api_get_rain}
\title{Get Rain Data from the Postgres Database Via API}
\usage{
api_get_rain(user_id, spot_id)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}
}
\description{
This function calls \code{\link{api_get_timeseries}} to get rain data records
from the "rains" endpoint and calls
\code{fhpredict:::remove_and_reorder_columns} to order the records
chronologically and to select only the relevant columns
}
