% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_timeseries.R
\name{api_get_timeseries}
\alias{api_get_timeseries}
\title{Get Timeseries Data from the Postgres Database Via API}
\usage{
api_get_timeseries(path, subject = "timeseries", sort = TRUE)
}
\arguments{
\item{path}{(relative) path to API endpoint}

\item{subject}{name of data subject, to be used in messages}

\item{sort}{if \code{TRUE} (the default), the returned data frame will be
sorted by the "dateTime" column}
}
\description{
Get Timeseries Data from the Postgres Database Via API
}
\keyword{internal}
