% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_user.R
\name{api_get_users}
\alias{api_get_users}
\title{Get Overview on Available Users}
\usage{
api_get_users()
}
\value{
data frame with one row per user and variables describing the user,
  such as: \code{id}, \code{firstName}, \code{lastName}, \code{rold},
  \code{email}
}
\description{
Get Overview on Available Users
}
