% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_rain_data_for_bathing_spot.R
\name{api_replace_rain}
\alias{api_replace_rain}
\title{Add Rain Data Replacing Records for Existing Times}
\usage{
api_replace_rain(
  user_id,
  spot_id,
  rain,
  rain_db = NULL,
  time_string,
  comment = ""
)
}
\arguments{
\item{user_id}{user identifier}

\item{spot_id}{bathing spot identifier}

\item{rain}{data frame with new rain data}

\item{rain_db}{optional. Data frame with existing rain data}

\item{time_string}{passed to \code{\link{api_add_rain}}}

\item{comment}{passed to \code{\link{api_add_rain}}}
}
\description{
This function reads existing rain data with \code{\link{api_get_rain}}
}
