% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_radolan.R
\name{download_radolan}
\alias{download_radolan}
\title{Download Radolan Files on DWD Server}
\usage{
download_radolan(temporal_resolution = "daily", export_dir = "data")
}
\arguments{
\item{temporal_resolution}{"daily" or "hourly" (default: daily)}

\item{export_dir}{export directory (default: "data" in current working
directory)}
}
\value{
list with "daily_historical_urls" and "hourly_historical_urls"
}
\description{
Download Radolan Files on DWD Server
}
