% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_area_coordinates.R
\name{get_area_coordinates}
\alias{get_area_coordinates}
\title{Return the coordinates of area assigned to bathing spot}
\usage{
get_area_coordinates(bathing_spot)
}
\arguments{
\item{bathing_spot}{list as returned by the API for one bathing spot}
}
\value{
matrix with two columns and as many rows as there are points
  describing the area
}
\description{
Return the coordinates of area assigned to bathing spot
}
