% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_summary.R
\name{get_data_summary}
\alias{get_data_summary}
\title{Get Overview on Available Data}
\usage{
get_data_summary(user_id, spot_id)
}
\arguments{
\item{user_id}{user identifier (see \code{\link{api_get_users}} for available
users)}

\item{spot_id}{bathing spot identifier (see
\code{\link{api_get_bathingspot}} for available bathing spots}
}
\value{
data frame with one row per type of data and columns \code{from}
  (date of earliest record), \code{to} (date of latest record), \code{n}
  (number of records)
}
\description{
Get Overview on Available Data
}
