% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_postgres_api_token.R
\name{get_postgres_api_token}
\alias{get_postgres_api_token}
\title{Get a Valid Token to Access the Postgres API}
\usage{
get_postgres_api_token(dbg = FALSE)
}
\arguments{
\item{dbg}{if \code{TRUE}, debug messages are shown}
}
\description{
If a token is stored locally it is checked whether this token is valid. If
the token is valid, it is returned. If the token is not valid or if there is
no token stored locally, request a new token, store it locally and return it.
}
