% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radolan.R
\name{get_radolan_urls}
\alias{get_radolan_urls}
\title{Get URLs to Available Radolan Files on DWD Server}
\usage{
get_radolan_urls()
}
\value{
list with "daily_historical_urls" and "hourly_historical_urls"
}
\description{
Get URLs to Available Radolan Files on DWD Server
}
