% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radolan_urls_bucket.R
\name{get_radolan_urls_bucket}
\alias{get_radolan_urls_bucket}
\title{Get URLs to Binary Radolan Filed From Flusshygiene's Amazon Bucket}
\usage{
get_radolan_urls_bucket(from = format(Sys.Date() - 1, "\%Y\%m\%d"),
  to = from, time = "")
}
\arguments{
\item{from}{first day considered in format yyyymmdd, e.g. "20190625".
Default: string representing "yesterday"}

\item{to}{last day considered in format yyyymmdd, e.g. "20190625". Default:
same day as given in \code{from}}

\item{time}{considered, given in format HHMM, e.g. "1050". Default: "", i.e.
URLs are not filtered by time}
}
\description{
This function requires an environment variable "TOKEN_PROD" to be set to the
token that is required to access the "Production Endpoint". Use
\code{usethis::edit_r_environ()} to open the .Renviron file in your home
directory and add a line "TOKEN_PROD=..." where ... is replaced with the
token.
}
\examples{
\dontrun{
# Get URLs to available files for 2019-06-25, 10:50
urls <- get_radolan_urls_bucket(from = "20190625", time = "1050")

# Read the first file (there should be only one such file!)
raster_rain <- kwb.dwd::read_binary_radolan_file(urls[1])
}

}
