% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radolan_urls_for_measurements.R
\name{get_radolan_urls_for_measurements}
\alias{get_radolan_urls_for_measurements}
\title{Determine URLs to Radolan Files for Measurements}
\usage{
get_radolan_urls_for_measurements(user_id, spot_id,
  sampling_time = "1050", date_range = NULL, all_in_range = FALSE,
  n_days_before = 5, summer_filter = FALSE)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{sampling_time}{expected sampling time. Will be used to select the
corresponding Radolan files.}

\item{date_range}{vector of two Date objects giving the first and last day of
rain data to be loaded. If \code{NULL} (the default) the range of dates is
determined from the range of dates for which water quality measurements are
available.}

\item{all_in_range}{if \code{TRUE} (the default is \code{FALSE}) rain data
for all days between the first and last day of a measurement are loaded.
By default (\code{all_in_range = FALSE}), data for the days of measurement
and for the days within a \code{n_days_before}-day time period before each
day of measurement, are loaded.}

\item{n_days_before}{number of days before a measurement for which to load
rain data}

\item{summer_filter}{if \code{TRUE}, only dates within the bathing season
are kept. Default: \code{FALSE}}
}
\value{
named vector of character containing the URLs to the Radolan files
  required for the measurements storted in the database. The names of the
  elements are strings in which the date and time is encoded in the format
  yyyymmddHHMM, e.g. "201809171050" for "2019-09-18 10:50"
}
\description{
Determine the URLs to the Radolan files that are required to calibrate
a model. For each day of measurement files (for the day of measurements and
the five days before) are required.
}
\examples{
\dontrun{
urls <- get_radolan_urls_for_measurements(user_id = 5, spot_id = 41)
}
}
