% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_validate_model.R
\name{get_stat_tests}
\alias{get_stat_tests}
\title{Calculate Statistical Tests for Residuals}
\usage{
get_stat_tests(model)
}
\arguments{
\item{model}{model object}
}
\description{
Normality and s2 = const shapiro-wilk test and breusch-pagan test
}
