% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_date_ranges.R
\name{group_dates_by_diff}
\alias{group_dates_by_diff}
\title{Build Groups of Dates by Difference}
\usage{
group_dates_by_diff(dates, diff_days = 1)
}
\arguments{
\item{dates}{vector of \code{Date} objects}

\item{diff_days}{maximum difference between consecutive dates that may occur
for dates within the same group}
}
\description{
Build Groups of Dates by Difference
}
\examples{
dates <- as.Date(c("2019-07-01", "2019-07-03", "2019-07-06", "2019-07-08"))
fhpredict:::group_dates_by_diff(dates, diff_days = 2)
}
