% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_paths.R
\name{path_bathingspot}
\alias{path_bathingspot}
\title{Path to Bathing Spot API endpoint}
\usage{
path_bathingspot(user_id = -1L, spot_id = -1L, limit = NULL,
  skip = NULL, sep = "/")
}
\description{
Path to Bathing Spot API endpoint
}
\examples{
fhpredict:::path_bathingspot()
fhpredict:::path_bathingspot(spot_id = 3)
fhpredict:::path_bathingspot(user_id = 7)
fhpredict:::path_bathingspot(user_id = 22, spot_id = 33)
}
\keyword{internal}
