% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_contigency.R
\name{plot_contigency}
\alias{plot_contigency}
\title{Model Classification Versus Percentiles}
\usage{
plot_contigency(model)
}
\arguments{
\item{model}{model object as returned by \code{\link{api_get_model}}}
}
\description{
Abbildung zeigt durch Model klassifizierte Werte im Vergelich zu den auf
diesen Werten berechneten Perzentilgrenzen
}
