% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_overview.R
\name{plot_data_overview}
\alias{plot_data_overview}
\title{Abbildung Datenqualität und Kalibrationsbereich}
\usage{
plot_data_overview(model)
}
\arguments{
\item{model}{model object as returned by \code{\link{api_get_model}}}
}
\description{
Diese Abbildung zeigt die Anzahl Regenassoziierter Datenpunkte
}
