% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_effects.R
\name{plot_effects}
\alias{plot_effects}
\title{Modellimplikationen}
\usage{
plot_effects(model)
}
\arguments{
\item{model}{model object as returned by \code{\link{api_get_model}}}
}
\description{
Modellimplikationen
}
