% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_vs_data.R
\name{plot_fit_vs_data}
\alias{plot_fit_vs_data}
\title{Plot Modelfit gegen Messwerte}
\usage{
plot_fit_vs_data(model)
}
\arguments{
\item{model}{model object as returned by \code{\link{api_get_model}}}
}
\description{
Abbildung zeigt Modelfit gegen Messwerte und 1 zu 1 Winkelhalbierende
}
