% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_model_calibration.R
\name{plot_model_calibration}
\alias{plot_model_calibration}
\title{Ueberblick ueber getestete Modelle}
\usage{
plot_model_calibration(tests, ref_model = "model_01")
}
\arguments{
\item{tests}{data frame similar to element \code{sorted_models} of the list
returned by \code{fhpredict:::build_and_validate_model}}

\item{ref_model}{name of the selected model (expected to be contained in
column \code{model} of \code{tests})}
}
\description{
Diese Funktion stellt die getestetet Modelle grafisch dar.
}
