% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres_delete}
\alias{postgres_delete}
\title{Send DELETE Request to Postgres API}
\usage{
postgres_delete(path, ...)
}
\arguments{
\item{path}{relative path, e.g. \code{users/3/bathingspots/18/genericInputs}
to delete a generic input for bathing spot with id 18 of user with id 3}

\item{\dots}{further arguments passed to \code{fhpredict:::postgres_request}}
}
\value{
In case of success this function returns what
  \code{httr::content(response, as = "parsed")} returns. In case of failure
  an empty list with attribute \code{response} (containing the response
  object returned by \code{\link[httr]{DELETE}}) is returned.
}
\description{
Send DELETE Request to Postgres API
}
