% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres.R
\name{postgres_get}
\alias{postgres_get}
\title{Send GET Request to Postgres API}
\usage{
postgres_get(path)
}
\arguments{
\item{path}{relative path, e.g. \code{bathingspots/1} to get data for
bathingspot with ID 1}
}
\value{
In case of success this function returns what
  \code{httr::content(response, "parsed")} returns. In case of failure an
  empty list with attribute \code{response} (containing the response object
  returned by \code{\link[httr]{GET}}) is returned.
}
\description{
Send GET Request to Postgres API
}
