% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres_get.R
\name{postgres_post}
\alias{postgres_post}
\title{Send POST Request to Postgres API}
\usage{
postgres_post(path, body = NULL)
}
\arguments{
\item{path}{relative path, e.g. \code{users/3/bathingspots/18/genericInputs}
to create a generic input for bathing spot of user 3 with ID 18}

\item{body}{list with the fields to be set as \code{key = value} pairs, e.g.
\code{list(name = "lirum larum")}}
}
\value{
In case of success this function returns what
  \code{httr::content(response, "parsed")$data} returns. In case of failure
  an empty list with attribute \code{response} (containing the response
  object returned by \code{\link[httr]{POST}}) is returned.
}
\description{
Send POST Request to Postgres API
}
