% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_token.R
\name{request_token}
\alias{request_token}
\title{Request a Token to Access the Flusshygiene Postgres API}
\usage{
request_token()
}
\description{
The function requires the following envirionment variables to be set:
\itemize{
  \item AUTH0_REQ_URL
  \item AUTH0_CLIENT_ID
  \item AUTH0_CLIENT_SECRET
  \item AUTH0_AUDIENCE
}
}
