% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fit_vs_data.R
\name{simulate_classification}
\alias{simulate_classification}
\title{Hilfsfunktion zum Vorhersagen generieren}
\usage{
simulate_classification(model)
}
\arguments{
\item{model}{model object as returned by \code{\link{api_get_model}}}
}
\description{
Hilfsfunktion zum Vorhersagen generieren
}
