% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleep.R
\name{sleep}
\alias{sleep}
\title{Wait a Given Number of Seconds}
\usage{
sleep(seconds)
}
\arguments{
\item{seconds}{time to wait, in seconds}
}
\value{
list with elements \code{data}, \code{success}, \code{message}
}
\description{
Wait a Given Number of Seconds
}
