% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_model.R
\name{api_add_model}
\alias{api_add_model}
\title{Add a Model to the Database}
\usage{
api_add_model(user_id, spot_id, model, comment = "any comment?")
}
\arguments{
\item{user_id}{user ID}

\item{spot_id}{bathing spot ID}

\item{model}{model object}

\item{comment}{description of the model. Default: "any comment?"}
}
\value{
This function returns the ID of the added model. The URL to the
  uploaded binary file is returned in attribute "model_url". From there,
  the model can be read back with \code{\link{readRDS}}.
}
\description{
Add a Model to the Database
}
