% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_get_timeseries.R
\name{api_get_timeseries}
\alias{api_get_timeseries}
\title{Get Timeseries Data from the Postgres Database Via API}
\usage{
api_get_timeseries(path, subject = "timeseries")
}
\arguments{
\item{path}{(relative) path to API endpoint}

\item{subject}{name of data subject, to be used in messages}
}
\description{
Get Timeseries Data from the Postgres Database Via API
}
\keyword{internal}
