% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_model}
\alias{build_model}
\title{Build Water Quality Model for Bathing Spot}
\usage{
build_model(user_id, spot_id, seed = NULL)
}
\arguments{
\item{user_id}{user ID}

\item{spot_id}{bathing spot ID}

\item{seed}{if \code{TRUE} the random number generate is initialised using
this value in a call to \code{\link{set.seed}}}
}
\value{
list with elements \code{data}, \code{success}, \code{message}
}
\description{
Build Water Quality Model for Bathing Spot
}
