% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_rain_data_for_bathing_spot.R
\name{get_unique_measurement_dates}
\alias{get_unique_measurement_dates}
\title{Sorted Unique Dates of Measurments}
\usage{
get_unique_measurement_dates(user_id, spot_id)
}
\description{
Sorted Unique Dates of Measurments
}
