% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quality.R
\name{predict_quality}
\alias{predict_quality}
\title{Predict Water Quality for Bathing Spot Using Stored Model}
\usage{
predict_quality(user_id, spot_id, date = Sys.Date())
}
\arguments{
\item{user_id}{user ID}

\item{spot_id}{bathing spot ID}

\item{date}{Date object or date string in format yyyy-mm-dd}
}
\value{
list with elements \code{data}, \code{success}, \code{message}
}
\description{
Predict Water Quality for Bathing Spot Using Stored Model
}
