% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_quality.R
\name{predict_quality}
\alias{predict_quality}
\title{Predict Water Quality for Bathing Spot Using Stored Model}
\usage{
predict_quality(
  user_id,
  spot_id,
  from = Sys.Date() - 1L,
  to = Sys.Date() + 1L,
  import = TRUE,
  return_debug_info = FALSE
)
}
\arguments{
\item{user_id}{user ID}

\item{spot_id}{bathing spot ID}

\item{from}{Date object or date string in format yyyy-mm-dd giving the first
day of the time period to be predicted. Default: "yesterday"}

\item{to}{Date object or date string in format yyyy-mm-dd giving the last day
of the time period to be predicted. Default: "tomorrrow"}

\item{import}{logical telling whether to import new rain data or not.
Default: \code{TRUE}.}

\item{return_debug_info}{logical with default \code{FALSE}. If \code{TRUE}
the prediction is not written to the database. Instead, what would be send
to the database is returned with all relevant variables that were used to
prepare the prediction being set as attributes.}
}
\value{
list with elements \code{data}, \code{success}, \code{message} or (if
  \code{return_debug_info = TRUE}) data frame representing the predictions
  with attributes \code{spot_data}, \code{riverdata_raw}, \code{riverdata},
  \code{newdata_raw}, \code{newdata}, \code{prediction} representing
  intermediate variables that were used to prepare the prediction  (see
  source code of \code{fhpredict::predict_quality} to understand their
  meaning)
}
\description{
If no time period is specified, the function will first check the
measurements for missing E.coli values and do the prediction for the days
for which an E.coli value of -1 has been uploaded. If all E.coli measurements
are valid, the prediction will be done for the current day
(\code{Sys.Date()}).
}
