% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_rain_data_for_bathing_spot.R
\name{provide_rain_data_for_bathing_spot}
\alias{provide_rain_data_for_bathing_spot}
\title{Provide Rain Data for one Bathing Spot}
\usage{
provide_rain_data_for_bathing_spot(
  user_id,
  spot_id,
  sampling_time = "1050",
  date_range = NULL
)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{sampling_time}{expected sampling time. Will be used to select the
corresponding Radolan files.}

\item{date_range}{vector of two Date objects giving the first and last day of
rain data to be loaded. If \code{NULL} (the default) the range of dates is
determined from the range of dates for which water quality measurements are
available.}
}
\value{
list with elements \code{data}, \code{success}, \code{message}. The
  \code{data} element contains a vector of IDs identifying the records in
  the rain data table associated with the bathing spot given by
  \code{user_id} and \code{spot_id}
}
\description{
Consider to use \code{\link{provide_rain_data}}. For a description see there.
}
\examples{
\dontrun{
provide_rain_data_for_bathing_spot(user_id = 5, spot_id = 41)
}
}
