% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/provide_rain_data_for_bathing_spot.R
\name{provide_rain_data_for_bathing_spot}
\alias{provide_rain_data_for_bathing_spot}
\title{Provide Rain Data for one Bathing Spot}
\usage{
provide_rain_data_for_bathing_spot(user_id, spot_id,
  sampling_time = "1050", date_range = NULL)
}
\arguments{
\item{user_id}{user id}

\item{spot_id}{bathing spot id}

\item{sampling_time}{expected sampling time. Will be used to select the
corresponding Radolan files.}

\item{date_range}{vector of two Date objects giving the first and last day of
rain data to be loaded. If \code{NULL} (the default) the range of dates is
determined from the range of dates for which water quality measurements are
available.}
}
\value{
vector of integer containing the IDs of the records inserted into the
  "rains" database table.
}
\description{
The function considers (i) the "catchment" area that is defined for the
bathing spot and (ii) the date range that is taken from the range of dates
for which measurements are available. Rain data are read from binary Radolan
files, cropped around the "catchment" area and and averaged over all layears
for each raster cell.
}
