% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{A io function}
\usage{
simple(spot_id, user_id)
}
\arguments{
\item{spot_id}{number}

\item{user_id}{number}
}
\description{
simple function to test the opencpu API for the project
}
\examples{
\dontrun{simple()}
}
\keyword{simple}
