% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{A io function}
\usage{
simple(str = "{\\"foo\\":\\"bah\\"}")
}
\arguments{
\item{str}{A JSON string (optional)}
}
\description{
simple function to test the opencpu API for the project
}
\examples{
\dontrun{simple()}
}
\keyword{simple}
