% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.R
\name{simple}
\alias{simple}
\title{A io function}
\usage{
simple(a = "foo", b = "bah")
}
\arguments{
\item{a}{string (optional)}

\item{b}{string (optional)}
}
\description{
simple function to test the opencpu API for the project
}
\examples{
\dontrun{simple()}
}
\keyword{simple}
